Refactor CLI api-keys command
